docker_scripts/export_poneglyph_dataset/requirements.txt
supabase
requests
pillow
tqdm
scikit-learn
python-dotenv
huggingface-hub
datasets
pillow-avif-plugin
